releasing version 2.0.5
git-svn-id: https://plugins.svn.wordpress.org/secure-wordpress/trunk@437306 b8457f37-d9ea-0310-8a92-e5e31aec5664

+++ uninstall.php
@@ -1,10 +1,12 @@
-<?php
-if( !defined( 'ABSPATH') && !defined('WP_UNINSTALL_PLUGIN') )
-	exit();
-
-delete_option('WSD-COOKIE');
-delete_option('WSD-TOKEN');
-delete_option('WSD-TARGETID');
-delete_option('WSD-USER');
+<?php if( !defined( 'ABSPATH') && !defined('WP_UNINSTALL_PLUGIN') ) { exit(); }
+/*
+ * Delete options
+ */
 delete_option('secure-wp');
-delete_option('wsd_feed_data');
+
+// this is a shared option so if there are
+// any other plug-ins installed, then don't
+// delete it.
+$__1 = ABSPATH.'wp-content/plugins/websitedefender-wordpress-security';
+$__2 = ABSPATH.'wp-content/plugins/wp-security-scan';
+if (!is_dir($__1) || !is_dir($__2)) { delete_option('wsd_feed_data'); }
